+++ src/energy_systems/connections/bus.jl
@@ -108,8 +108,8 @@ function balance_nr(unit::Bus, caller::Bus)::Float64
         end
 
         if isa(inface.source, Bus)  
-            InterfaceInfo = balance_nr(inface.source, unit)
-            balance_supply = max(InterfaceInfo, inface.balance)
+            exchange = balance_nr(inface.source, unit)
+            balance_supply = max(exchange, inface.balance)
             if balance_supply < 0.0
                 continue
             end
@@ -125,8 +125,8 @@ function balance_nr(unit::Bus, caller::Bus)::Float64
         end
 
         if isa(outface.target, Bus)
-            InterfaceInfo = balance_nr(outface.target, unit)
-            balance_demand = min(InterfaceInfo, outface.balance)
+            exchange = balance_nr(outface.target, unit)
+            balance_demand = min(exchange, outface.balance)
             if balance_demand > 0.0
                 continue
             end
@@ -213,11 +213,11 @@ function balance_on(
     # iterate through outfaces to get storage loading potential
     for (idx, outface) in pairs(unit.output_interfaces)
         if outface.target.sys_function === sf_bus
-            InterfaceInfo = balance_on(outface, outface.target)
-            balance = InterfaceInfo.balance
-            storage_potential = InterfaceInfo.storage_potential
-            energy_potential = outface.sum_abs_change > 0.0 ? 0.0 : InterfaceInfo.energy_potential
-            temperature = InterfaceInfo.temperature
+            exchange = balance_on(outface, outface.target)
+            balance = exchange.balance
+            storage_potential = exchange.storage_potential
+            energy_potential = outface.sum_abs_change > 0.0 ? 0.0 : exchange.energy_potential
+            temperature = exchange.temperature
         else
             balance = outface.balance
             temperature = outface.temperature
@@ -233,8 +233,8 @@ function balance_on(
                     || unit.connectivity.storage_loading[input_index][get_connectivity_output_index(unit, idx)]
                 )
             )
-                InterfaceInfo = balance_on(outface, outface.target)
-                storage_potential = InterfaceInfo.storage_potential
+                exchange = balance_on(outface, outface.target)
+                storage_potential = exchange.storage_potential
             else
                 storage_potential = 0.0
             end
@@ -253,8 +253,8 @@ function balance_on(
     if caller_is_input == false && interface.sum_abs_change == 0.0 # also need to check inputs of unit in order to sum up potential_energy_inputs, but only if necessary
         for (idx, inface) in pairs(unit.input_interfaces)
             if inface.source.sys_function === sf_bus
-                InterfaceInfo = balance_on(inface, inface.source)
-                energy_potential = inface.sum_abs_change > 0.0 ? 0.0 : InterfaceInfo.energy_potential
+                exchange = balance_on(inface, inface.source)
+                energy_potential = inface.sum_abs_change > 0.0 ? 0.0 : exchange.energy_potential
             else
                 energy_potential = (inface.max_energy === nothing || inface.sum_abs_change > 0.0 ) ? 0.0 : inface.max_energy
             end
@@ -270,11 +270,11 @@ function balance_on(
     #       If an energy system connected to the interface of balane_on() has already been produces, the 
     #       max_energy is ignored and set to zero by balance_on(). Then, only the balance can be used in the 
     #       calling energy system to avoid double counting.
-    
+    balance_written = interface.sum_abs_change > 0.0
     return (
             balance = balance(unit),
             storage_potential = caller_is_input ? storage_potential_outputs : 0.0,
-            energy_potential = interface.sum_abs_change > 0.0 ? 0.0 : (caller_is_input ? energy_potential_outputs : energy_potential_inputs) ,
+            energy_potential = balance_written ? 0.0 : (caller_is_input ? -abs(energy_potential_outputs) : abs(energy_potential_inputs)),
             temperature = (highest_demand_temp <= -1e9 ? nothing : highest_demand_temp)
             )
 end