+++ test/order_of_operations/bus_output_priorities.jl
@@ -106,7 +106,6 @@ function test_ooo_bus_output_priorities()
         ("TST_BUS_02", EnergySystems.s_produce),
         ("TST_BUS_01", EnergySystems.s_produce),
         ("TST_BUS_03", EnergySystems.s_produce),
-        ("TST_GBO_01", EnergySystems.s_potential),
         ("TST_GBO_01", EnergySystems.s_produce),
         ("TST_GRI_01", EnergySystems.s_produce),
         ("TST_BUS_02", EnergySystems.s_distribute),